feat: alerts expand the form instead of after the form
